updated UI for club application list item
